+++ reapy/reascript_api/dist_api.py
@@ -1,7 +1,0 @@
-import reapy
-if not reapy.is_inside_reaper():
-    try:
-        from reapy_generated_api import __all__
-        from reapy_generated_api import *
-    except ImportError:  # Happens when ``reapy`` dist API is disabled
-        __all__ = []